Add tuple parameters to execute
The automatic promotion of array elements in Julia can lead to bugs when executing
stored queries with LibPQ.jl. For example, when inserting a row with both integer
and double precision columns, the current API of LibPQ.jl requires you to do
something like `execute(stmt, [1, 2.0])`, but the parameters will be promoted by
Julia to `[1.0, 2.0]`, causing an error. The way to avoid this catch with the
current API is to pass the parameters as `Real[1, 2.0]` or `Any[1, 2.0]`.

This commit allows parameters to `execute` to be a `Tuple`, which avoids these
pitfalls with automatic promotion.