Allow concrete array types to be specified with column_types
Previously, it was not possible to set a column type to something like
`Vector{Int32}`, and one would instead have to specify `AbstractArray{Int32}`.
By changing the parse method for arrays to accept all subtypes of
`AbstractArray`, it is now possible to specify `Vector{Int32}`, or any other
concrete type of array, as a column type.